Creado componente "Counter". Implementados contadores de tareas en App y confirmación de eliminación de tareas en TaskItem.

+++ src/components/TaskItem.jsx
@@ -1,18 +1,28 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import PropTypes from "prop-types";
 import { BiTask, BiTaskX } from "react-icons/bi";
 
 const TaskItem = ({ task, handleCompleteTask, handleDeleteTask }) => {
   const [completed, setCompleted] = useState(task.completed);
+  const [confirmDelete, setConfirmDelete] = useState(false);
+
+  const confirmDeletion = () => {
+    const isConfirmed = window.confirm(
+      "¿Seguro que deseás eliminar esta tarea?"
+    );
+    setConfirmDelete(isConfirmed); // Pedir confimación al usuario para setear el estado confirmDelete.
+  };
 
   const handleComplete = () => {
     setCompleted(!completed);
-    handleCompleteTask(task.id); // Llama a la función de App.jsx para actualizar el estado
+    handleCompleteTask(task.id); // Llamar a la función de App.jsx para actualizar el estado.
   };
 
-  const handleDelete = () => {
-    handleDeleteTask(task.id); // Llama a la función de App.jsx para eliminar la tarea
-  };
+  useEffect(() => {
+    if (confirmDelete) {
+      handleDeleteTask(task.id); // Llamar a la función de App.jsx para eliminar la tarea solo si se ha confirmado por el usuario.
+    }
+  }, [confirmDelete, handleDeleteTask, task.id]);
 
   return (
     <div style={{ textDecoration: completed ? "line-through" : "none" }}>
@@ -23,7 +33,7 @@ const TaskItem = ({ task, handleCompleteTask, handleDeleteTask }) => {
         onClick={handleComplete}>
         <BiTask />
       </button>
-      <button className="btn" title="Eliminar tarea" onClick={handleDelete}>
+      <button className="btn" title="Eliminar tarea" onClick={confirmDeletion}>
         <BiTaskX />
       </button>
     </div>
